Adding general projects page to website. added general:boolean column to model data_types.
Summary: We will be adding a general projects page to our web application. The general page itself will be an individualized project that contains data types and values that are only specific to general overview of projects. Therefore, in the process of creating general projects we will also add a column in data types model that determines whether our data type is for the general project or not.

Test Plan: Open rails server and go to general projects page to view it.

Reviewers: sam, jodreen

Reviewed By: jodreen

Subscribers: sam, kwu, isaiahyoon

Projects: #crt

Differential Revision: http://phab.calblueprint.org/D65