Adding order to data_types. Editable on forms.
Summary: Data types have order attribute that determines how they are displayed to the user. This includes the ordering in Projects Data page and the ordering in index page of data types. Order can be edited on the data type forms.

Test Plan: Go to data types page. Click on a data type and update the value of order to new value. After clicking update, check to see that the order of the two data types that have switched orders have switched positions. Next, go to the Projects page and view the project data of any project. Verify that the data for those corresponding data types have switched rows.

Reviewers: jodreen, jsonderulo

Reviewed By: jsonderulo

Subscribers: jsonderulo, kwu, isaiahyoon

Projects: #crt

Differential Revision: http://phab.calblueprint.org/D236