Updating all associated data values after one edit is made in Projects data table.
Summary: Once and edit is made in the Project data table, every associated data type that has a dependency on this value is updated as well within the Projects data table. Thus, to veiw updated data, the user will not need to refresh the page each time an edit is made.

Test Plan: Go to a specific project. Edit a data value that is depended on by another data types formula. Ensure proper values are edited after update request is fired.

Reviewers: jodreen, jsonderulo

Reviewed By: jsonderulo

Subscribers: jsonderulo, kwu, isaiahyoon

Projects: #crt

Differential Revision: http://phab.calblueprint.org/D286