Fix the overreaching JS from the Excel hotkeys
Summary:
The old JS prevented the default behavior for every space bar or enter
keydown. This branch only does that if you're on the project page, as it
should be.

Test Plan:
Ensure that the Excel hotkeys still work, while testing on other pages
if you can hit Enter and have it work.

Reviewers: jsonderulo, isaiahyoon

Reviewed By: isaiahyoon

Subscribers: jonchuchu, quinton, qin

Projects: #crt

Differential Revision: http://phab.calblueprint.org/D444